refactor(catalog): complete song controller dto refactoring, add checking if resource exists when getting by reference

+++ src/main/java/com/spring/beatmarket/domain/catalog/dto/song/UpdateSongDto.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public record UpdateSongDto(
         Optional<String> title,
         Optional<LocalDate> releaseDate,
-        Optional<Long> duration,
+        Optional<Integer> duration,
         Optional<SongLanguage> language,
         Optional<Long> genreId,
         Optional<List<Long>> artistIds,
